fix(all) : 2023. 10 Survivor 일차적인 스크립트 완성
미니맵을 이용해 명령 전달 시 cmdType이 없어 명령 수행이 안되는 현상이나
스크립터블 오브젝트에 동적 할당된 유닛의 정보를 담으려고 하니
typemismatch로 확인이 안되는 현상 같이 자잘한 버그 해결

이제 건물 종류나, 적과 아군 유닛 배치 그리고 지형 제작하면 게임이
완성된다

+++ 0_Scripts/Project/Survival/00_Select/0_Base/Command.cs
@@ -18,13 +18,13 @@ public class Command
 
     public STATE_SELECTABLE type;
 
-    public Command(ushort _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
+    public Command(int _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
     {
 
         Set(_unitNums, _type, _pos, _target);
     }
 
-    public Command(ushort _unitNums, STATE_SELECTABLE _type)
+    public Command(int _unitNums, STATE_SELECTABLE _type)
     {
 
         Set(_unitNums, _type);
@@ -65,10 +65,10 @@ public class Command
         ChkAllReceived();
     }
 
-    public void Set(ushort _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
+    public void Set(int _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
     {
 
-        unitNums = _unitNums;
+        unitNums = (ushort)_unitNums;
         receiveNum = 0;
         type = _type;
 
@@ -76,10 +76,10 @@ public class Command
         pos = _pos;
     }
 
-    public void Set(ushort _unitNums, STATE_SELECTABLE _type)
+    public void Set(int _unitNums, STATE_SELECTABLE _type)
     {
 
-        unitNums = _unitNums;
+        unitNums = (ushort)_unitNums;
         receiveNum = 0;
         type = _type;
 
@@ -134,7 +134,7 @@ public class Command
         }
     }
 
-    public static Command GetCommand(ushort _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
+    public static Command GetCommand(int _unitNums, STATE_SELECTABLE _type, Vector3 _pos, Selectable _target = null)
     {
 
         if (pool == null) pool = new Stack<Command>(VariableManager.MAX_SAVE_COMMANDS);
@@ -154,7 +154,7 @@ public class Command
         else return new Command(_unitNums, _type, _pos, _target);
     }
 
-    public static Command GetCommand(ushort _unitNums, STATE_SELECTABLE _type)
+    public static Command GetCommand(int _unitNums, STATE_SELECTABLE _type)
     {
 
         if (pool == null) pool = new Stack<Command>(VariableManager.MAX_SAVE_COMMANDS);
